+++ drivers/mysql/mysql.go
@@ -20,7 +20,7 @@ import (
 	"github.com/go-sql-driver/mysql"
 	"github.com/xo/dburl"
 
-	"github.com/neilotoole/sq/libsq/ast/sqlbuilder"
+	"github.com/neilotoole/sq/libsq/ast/render"
 	"github.com/neilotoole/sq/libsq/core/errz"
 	"github.com/neilotoole/sq/libsq/core/kind"
 	"github.com/neilotoole/sq/libsq/core/sqlmodel"
@@ -93,8 +93,8 @@ func placeholders(numCols, numRows int) string {
 }
 
 // SQLBuilder implements driver.SQLDriver.
-func (d *driveri) Renderer() *sqlbuilder.Renderer {
-	r := sqlbuilder.NewDefaultRenderer()
+func (d *driveri) Renderer() *render.Renderer {
+	r := render.NewDefaultRenderer()
 	return r
 }
