+++ drivers/sqlite3/sqlite3.go
@@ -26,7 +26,7 @@ import (
 	"golang.org/x/exp/slog"
 
 	_ "github.com/mattn/go-sqlite3" // Import for side effect of loading the driver
-	"github.com/neilotoole/sq/libsq/ast/sqlbuilder"
+	"github.com/neilotoole/sq/libsq/ast/render"
 	"github.com/neilotoole/sq/libsq/core/errz"
 	"github.com/neilotoole/sq/libsq/core/kind"
 	"github.com/neilotoole/sq/libsq/core/sqlmodel"
@@ -182,8 +182,8 @@ func placeholders(numCols, numRows int) string {
 }
 
 // SQLBuilder implements driver.SQLDriver.
-func (d *driveri) Renderer() *sqlbuilder.Renderer {
-	r := sqlbuilder.NewDefaultRenderer()
+func (d *driveri) Renderer() *render.Renderer {
+	r := render.NewDefaultRenderer()
 	return r
 }
