+++ drivers/sqlserver/sqlserver.go
@@ -21,7 +21,7 @@ import (
 
 	mssql "github.com/microsoft/go-mssqldb"
 
-	"github.com/neilotoole/sq/libsq/ast/sqlbuilder"
+	"github.com/neilotoole/sq/libsq/ast/render"
 	"github.com/neilotoole/sq/libsq/core/errz"
 	"github.com/neilotoole/sq/libsq/core/kind"
 	"github.com/neilotoole/sq/libsq/core/sqlmodel"
@@ -108,8 +108,8 @@ func placeholders(numCols, numRows int) string {
 }
 
 // SQLBuilder implements driver.SQLDriver.
-func (d *driveri) Renderer() *sqlbuilder.Renderer {
-	r := sqlbuilder.NewDefaultRenderer()
+func (d *driveri) Renderer() *render.Renderer {
+	r := render.NewDefaultRenderer()
 
 	// Custom functions for SQLServer specific stuff.
 	r.Range = renderRange
