+++ libsq/ast/render/function.go
@@ -1,4 +1,4 @@
-package sqlbuilder
+package render
 
 import (
 	"strings"
