+++ libsq/ast/render/internal_test.go
@@ -1,4 +1,4 @@
-package sqlbuilder
+package render
 
 import (
 	"testing"
