+++ libsq/ast/render/selectcols.go
@@ -1,4 +1,4 @@
-package sqlbuilder
+package render
 
 import (
 	"fmt"
